Huge update: Added Cloning of Data sheet & pyomo model; added new file for declaration of alternative scenarios scenario.py; added saving of each pyomo model to .lp file
